feat: introduce voidresult
